refactor: add HelpTooltip, group tooltip icons
