+++ src/renderer/src/components/TooltipIcons/__tests__/InfoTooltip.test.tsx
