+++ src/renderer/src/components/TooltipIcons/index.ts
@@ -1,0 +1,3 @@
+export { default as HelpTooltip } from './HelpTooltip'
+export { default as InfoTooltip } from './InfoTooltip'
+export { default as WarnTooltip } from './WarnTooltip'
