+++ src/renderer/src/pages/memory/settings-modal.tsx
@@ -1,8 +1,8 @@
 import { loggerService } from '@logger'
 import AiProvider from '@renderer/aiCore'
-import InfoTooltip from '@renderer/components/InfoTooltip'
 import InputEmbeddingDimension from '@renderer/components/InputEmbeddingDimension'
 import ModelSelector from '@renderer/components/ModelSelector'
+import { InfoTooltip } from '@renderer/components/TooltipIcons'
 import { isEmbeddingModel, isRerankModel } from '@renderer/config/models'
 import { useModel } from '@renderer/hooks/useModel'
 import { useProviders } from '@renderer/hooks/useProvider'
