+++ src/renderer/src/pages/settings/DocProcessSettings/OcrSystemSettings.tsx
@@ -1,6 +1,6 @@
 // import { loggerService } from '@logger'
-import InfoTooltip from '@renderer/components/InfoTooltip'
 import { SuccessTag } from '@renderer/components/Tags/SuccessTag'
+import { InfoTooltip } from '@renderer/components/TooltipIcons'
 import { isMac, isWin } from '@renderer/config/constant'
 import { useOcrProvider } from '@renderer/hooks/useOcrProvider'
 import useTranslate from '@renderer/hooks/useTranslate'
